@@ -3,14 +3,15 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_speech/google_speech.dart';
-import 'package:google_speech/speech_to_text_beta.dart';
 import 'package:path_provider/path_provider.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -20,12 +21,14 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: AudioRecognize(),
+      home: const AudioRecognize(),
     );
   }
 }
 
 class AudioRecognize extends StatefulWidget {
+  const AudioRecognize({Key? key}) : super(key: key);
+
   @override
   State<StatefulWidget> createState() => _AudioRecognizeState();
 }
@@ -40,7 +43,7 @@ class _AudioRecognizeState extends State<AudioRecognize> {
       recognizing = true;
     });
     final serviceAccount = ServiceAccount.fromString(
-        '${(await rootBundle.loadString('assets/test_service_account.json'))}');
+        (await rootBundle.loadString('assets/test_service_account.json')));
     final speechToText = SpeechToTextBeta.viaServiceAccount(serviceAccount);
     final config = _getConfig();
     final audio = await _getAudioContent('test.wav');
@@ -62,7 +65,7 @@ class _AudioRecognizeState extends State<AudioRecognize> {
       recognizing = true;
     });
     final serviceAccount = ServiceAccount.fromString(
-        '${(await rootBundle.loadString('assets/test_service_account.json'))}');
+        (await rootBundle.loadString('assets/test_service_account.json')));
     final speechToText = SpeechToTextBeta.viaServiceAccount(serviceAccount);
     final config = _getConfig();
 
@@ -88,8 +91,8 @@ class _AudioRecognizeState extends State<AudioRecognize> {
       model: RecognitionModel.basic,
       enableAutomaticPunctuation: true,
       sampleRateHertz: 16000,
-      languageCode: 'de-DE',
-      alternativeLanguageCodes: ['fr-FR', 'en-US']);
+      languageCode: 'en-US',
+      alternativeLanguageCodes: ['fr-FR', 'de-US']);
 
   Future<void> _copyFileFromAssets(String name) async {
     var data = await rootBundle.load('assets/$name');
@@ -121,7 +124,7 @@ class _AudioRecognizeState extends State<AudioRecognize> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Audio File Example'),
+        title: const Text('Audio File Example'),
       ),
       body: Center(
         child: Column(
@@ -131,20 +134,20 @@ class _AudioRecognizeState extends State<AudioRecognize> {
               _RecognizeContent(
                 text: text,
               ),
-            RaisedButton(
+            ElevatedButton(
               onPressed: recognizing ? () {} : recognize,
               child: recognizing
-                  ? CircularProgressIndicator()
-                  : Text('Test with recognize'),
+                  ? const CircularProgressIndicator()
+                  : const Text('Test with recognize'),
             ),
-            SizedBox(
+            const SizedBox(
               height: 10.0,
             ),
-            RaisedButton(
+            ElevatedButton(
               onPressed: recognizing ? () {} : streamingRecognize,
               child: recognizing
-                  ? CircularProgressIndicator()
-                  : Text('Test with streaming recognize'),
+                  ? const CircularProgressIndicator()
+                  : const Text('Test with streaming recognize'),
             ),
           ],
         ),
@@ -154,9 +157,9 @@ class _AudioRecognizeState extends State<AudioRecognize> {
 }
 
 class _RecognizeContent extends StatelessWidget {
-  final String text;
+  final String? text;
 
-  const _RecognizeContent({Key key, this.text}) : super(key: key);
+  const _RecognizeContent({Key? key, this.text}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -164,14 +167,14 @@ class _RecognizeContent extends StatelessWidget {
       padding: const EdgeInsets.all(16.0),
       child: Column(
         children: <Widget>[
-          Text(
+          const Text(
             'The text recognized by the Google Speech Api:',
           ),
-          SizedBox(
+          const SizedBox(
             height: 16.0,
           ),
           Text(
-            text,
+            text ?? '---',
             style: Theme.of(context).textTheme.bodyText1,
           ),
         ],
